Remove extra space in the "RPG game" link.
In the result view, there is a space before the "RPG game" link. It's not looking good, because the underline is shown in space, and it looks like "_RPG game". So I removed a space from the start of <a> tag and added a space at end of <p> tag, and It looks fine :)