Typescript conversion (#62)
* Full conversion to Typescript

All we have left to do is see if it works on GitHub.

* Keep trying to make sense of every type

Not all synthetic events are created equal

* Finish my conversion. No implict anys left

* Cleanup style and syntax

I also cleaned up things so that there were actual interfaces used. The validation was much needed. In addition I reformatted most of the files to match up to the standards set forth in the .eslintrc.json file; which can be altered. An very basic editorconfig was provided. The simple Function types were replaced with something easier to check.