Chore: Create tests for a few components
I wanted there to be a some tests set up because every single
app ideally should have more than the default tests set up.
As some components are intertwined , I am saving those for a
later date. I am interested in hearing further clarification
regarding the SelectQuiz component and if it would be wise
to edit it such that the last index is always the largest
amount of questions. Jest and React Testing Library are
currently powering the tests, further tools can be added
as required.