Chore: Create tests for a few components (#88)
* Chore: Create tests for a few components

I wanted there to be a some tests set up because every single
app ideally should have more than the default tests set up.
As some components are intertwined , I am saving those for a
later date. I am interested in hearing further clarification
regarding the SelectQuiz component and if it would be wise
to edit it such that the last index is always the largest
amount of questions. Jest and React Testing Library are
currently powering the tests, further tools can be added
as required.

* Realize typos and concerns

The wording is a lot about better thanks to the suggestions.
Additionally while it is a non-issue because of the question amount
we have; it didn't feel right to let 600 be the cut-off point for
all.

* Reorganize the tests

Okay this should be good now. Every component has it's own
test file and every test is under the __tests__ folder.

* Test folder not tests folder oops.